+++ src/main/java/com/sb11/hr_bank/domain/backup/service/BackupService.java
@@ -1,6 +1,6 @@
-package com.sb11.hr_bank.backup.service;
+package com.sb11.hr_bank.domain.backup.service;
 
-import com.sb11.hr_bank.backup.entity.Backup;
+import com.sb11.hr_bank.domain.backup.entity.Backup;
 import java.util.List;
 
 public interface BackupService {
